# openapi3_1__api_gateway.yaml
openapi: 3.1.0
info:
  title: Symcon smart-home — API Gateway
  description: >-
    API шлюза сервиса управления устройствами умного дома
    для взаимодействия с клиентскими приложениями.
  contact:
    name: Alexey Kuznetsov
    url: https://github.com/alexeycoder
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
servers:
  - url: https://localhost:8080/v2
    description: Test server
  - url: https://symcon.ru/v2
    description: Production server
...
paths:
...
  /devices/{deviceUuid}:

    get:
      tags:
        - devices
      summary: Получает данные устройства
      description: Получает данные устройства
      operationId: getDevice
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Некорректный идентификатор устройства
        '401':
          description: Несанкционированный доступ
        '404':
          description: Устройство не найдено
        default:
          description: Ошибка сервера

    put:
      tags:
        - devices
      summary: Обновляет данные устройства
      description: Обновляет данные устройства
      operationId: updateDevice
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Данные устройства для обновления
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
        required: true
      responses:
        '200':
          description: Данные устройства успешно обновлены
        '400':
          description: Некорректный идентификатор устройства
        '401':
          description: Несанкционированный доступ
        '404':
          description: Устройство не найдено
        default:
          description: Ошибка сервера

    delete:
      tags:
        - devices
      summary: Удаляет устройство
      description: Удаляет устройство
      operationId: deleteDevice
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Устройство успешно удалено
        '400':
          description: Некорректный идентификатор устройства
        '401':
          description: Несанкционированный доступ
        '404':
          description: Устройство не найдено
        default:
          description: Ошибка сервера
  /devices/{deviceUuid}/state:

    get:
      tags:
        - devices
      summary: Получает основные показатели состояния устройства
      description: |
        Получает основные показатели состояния устройства.

        При нахождении устройства и клиентского приложения в одном сегменте сети,
        приложению следует опрашивать устройство напрямую минуя API сервиса.
        
      operationId: getDeviceState
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceState'
        '400':
          description: Некорректные идентификатор устройства
        '401':
          description: Несанкционированный доступ
        '404':
          description: Устройство не найдено
        default:
          description: Ошибка сервера
  /devices/{deviceUuid}/schedules:

    get:
      tags:
        - schedules
      summary: Получает все графики работ устройства
      description: Получает все графики работ устройства
      operationId: findSchedules
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        '400':
          description: Некорректный идентификатор устройства
        '401':
          description: Несанкционированный доступ
        '404':
          description: Устройство не найдено
        default:
          description: Ошибка сервера

    post:
      tags:
        - schedules
      summary: Добавляет график работ для устройства
      description: Добавляет график работ для устройства
      operationId: addSchedule
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Данные графика работ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '201':
          description: График работ успешно добавлен
          headers:
            Location:
              description: URL добавленного нового графика работ
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: object
                properties:
                  scheduleNum:
                    type: integer
                    format: int32
                    description: Номер графика в разрезе устройства
          links:
            GetDeviceScheduleByNum:
              operationId: getSchedule
              parameters:
                scheduleNum: "$response.body#/scheduleNum"
              description: >-
                Значение `scheduleNum`, полученное в ответе, может использоваться
                в качестве одноимённого параметра в
                `GET /devices/{deviceUuid}/schedules/{scheduleNum}`.
            UpdateDeviceScheduleByNum:
              operationId: updateSchedule
              parameters:
                scheduleNum: "$response.body#/scheduleNum"
              description: >-
                Значение `scheduleNum`, полученное в ответе, может использоваться
                в качестве одноимённого параметра в
                `PUT /devices/{deviceUuid}/schedules/{scheduleNum}`.
        '400':
          description: >
            Некорректные данные графика работ или
            некорректный идентификатор устройства в запросе
        '401':
          description: Несанкционированный доступ
        '404':
          description: Устройство не найдено
        default:
          description: Ошибка сервера
  /devices/{deviceUuid}/schedules/{scheduleNum}:

    get:
      tags:
        - schedules
      summary: Получает график работ устройства по номеру
      description: Получает график работ устройства по номеру
      operationId: getSchedule
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleNum
          in: path
          description: Номер графика работ
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          description: Некорректные идентификатор устройства или номер графика
        '401':
          description: Несанкционированный доступ
        '404':
          description: График не найден
        default:
          description: Ошибка сервера

    put:
      tags:
        - schedules
      summary: Обновить график работ
      description: Обновить график работ
      operationId: updateSchedule
      security:
        - ApiTokenAuth: []
      parameters:
        - name: deviceUuid
          in: path
          description: Уникальный идентификатор устройства
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleNum
          in: path
          description: Номер графика работ
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Данные графика работ для обновления
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Некорректные идентификатор устройства или номер графика
        '401':
          description: Несанкционированный доступ
        '404':
          description: График не найден
        default:
          description: Ошибка сервера
components:
  securitySchemes:
    ApiTokenAuth:
      type: apiKey
      in: header
      name: Api-Token
  schemas:
...
    Device:
      required:
        - typeId
        - name
        - modelId
        - serialNum
        - firmwareVersion
        - connectionId
      type: object
      properties:
        deviceUuid:
          type: string
          format: uuid
          examples:
            - 'a83dcf43-dd10-4597-9a9c-6ba1ac50bd0c'
        typeId:
          type: integer
          format: int32
          examples:
            - 1
            - 2
          description: Код производителя типа устройства, например, 1 - робот-пылесос
        name:
          type: string
          examples:
            - Робот Кузя
        modelId:
          type: integer
          format: int32
        serialNum:
          type: string
          examples:
            - "6a0fa287-6969-4839"
        firmwareVersion:
          type: integer
          format: int32
          examples:
            - 2
        connectionId:
          type: integer
          format: int64
          examples:
            - 12
    DeviceState:
      description: |
        Основные показатели состояния устройства.
        API сервиса позволяет получить состояние устройства удалённо,
        когда устройство и клиентское приложение находятся в разных сегментах
        (разных сетях).
        В обычном случае, когда устройство и клиентское приложение находяся
        в одной локальной сети, получение данных о состоянии устройства должно
        осуществляться приложением напрямую в обход API сервиса.

      required:
        - deviceUuid
        - chargeLevel
        - isCharging
        - batteryHealth
        - currentService
        - connectionUrl
      type: object
      properties:
        deviceUuid:
          type: string
          format: uuid
          examples:
            - 'a83dcf43-dd10-4597-9a9c-6ba1ac50bd0c'
        chargeLevel:
          type: integer
          format: int32
          examples:
            - 0
            - 100
        isCharging:
          type: boolean
        batteryHealth:
          type: integer
          format: int32
          examples:
            - 80
        currentService:
          type: string
          format: enum
        connectionUrl:
          type: string
          format: uri
          examples:
            - "192.168.1.123:45678"
    Schedule:
      required:
        - deviceUuid
        - dayOfWeek
        - timeOfDay
        - serviceTypeId
        - servicePropertiesCode
      type: object
      properties:
        deviceUuid:
          type: string
          format: uuid
          examples:
            - 'a83dcf43-dd10-4597-9a9c-6ba1ac50bd0c'
        dayOfWeek:
          type: integer
          format: int32
        timeOfDay:
          type: string
          format: time
          examples:
            - "10:35"
        service:
          type: string
          enum:
            - RoboVac_Vacuum_Dry
            - RoboVac_Vacuum_Wet
            - RoboVac_EmptyDuster
            - RoboVac_CancelAndBack
            - RoboVac_BuildMap